requirements.txt
# BinBot Dependencies

# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Google Gemini AI
google-generativeai==0.3.2

# Vector database
chromadb==0.4.18

# Image processing
Pillow==10.1.0

# HTTP client for external APIs
httpx==0.25.2

# Environment variables
python-dotenv==1.0.0

# Data validation
pydantic==2.5.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # for testing FastAPI

# Development
black==23.11.0
flake8==6.1.0
